backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=24.1
pandas>=2.0
numpy>=1.24
statsmodels>=0.14
openpyxl>=3.1
